requirements.txt
h5py>=2.6
numpy>=1.11
py3helpers
PyYAML
six>=1.10
progressbar33>=2.3.1
ont-fast5-api
setuptools>=42.0.2
cython>=0.29.14
numpy>=1.18.0
scikit-learn>=0.22
h5py>=2.10.0
tqdm>=4.41.0
matplotlib>=3.0.3
seaborn>=0.9.0
biopython>=1.76
pysam>=0.15.3
pandas>=0.24.0
scipy>=1.3.0
